Make Contact#{closed?, open?} mirror Event.  #3221.
This commit also eliminates Contact#completed? and Contact#complete?
without supplying a replacement.

There is no meaningful definition of contact completion in Cases.  A
contact may be closed with a disposition code that indicates success
(e.g. as defined by Mdes::DispositionCode#success?), but it doesn't tell
you anything about the events done on that contact.  (To find that out,
you'd want to use Event#completed?.)

As far as I can tell, nothing used Contact#completed? anyway, so this
should be okay.