Removed new_event_for_person
For a long time this method has been marked as something
to be removed. During the audit of Time.now usage #2374
we again came across this method. It was called when neither
a Participant nor an Event were provided for a Contact.

This change will now raise an Exception in this instance. I have
created a new task #3285 to handle the situation when a Contact
is created outside of the context of an Event with a Person
who is not a Participant. This type of Contact should be possible.