Set participant.enroll_date #3288
Cases was not setting the enroll date on Participant when
the Participant was being enrolled. This commit sets the enroll
date to the ParticipantConsent.consent_date (if available) or today.

We also include a migration to set the enroll date to the date that it
should have been set to - either the first consent where consent was
given or the date of the record where consent was determined.

Closes #3288