Eliminate use of Field::Validator.expanded_schema in merge spec.
The expanded schema was used to figure out what type of values to
generate for the attribute merge tests.  This new approach takes
advantage of the fact that attribute names are standardized to Cases'
models (via the hash adapters), and so derives the type by introspecting
on the appropriate model class.

Since the schema is supposed to reflect what Cases' data model requires
and not the other way around, I think this change is fine.  Schema
errors will just have to be caught and fixed the old-fashioned way:
by reading the code.

This commit also alters test value generation for string fields that are
times.  Instead of generating random strings, it now randomly selects a
time formatted as HH:MM.