Click to finish link was missing #3319
When a participant was scheduled multiple times for the same event
in PSC, Cases did not take the date of the event. This was causing
the sections menu to show duplicate activities in the UI and was
not allowing the user to "finish" the survey thinking that the
survey was a "multi-part" survey.

This commit fixes this issue and uses the Event rather than just the
event type to determine the activities.

All instrument_plan methods that use just the event type name
instead of the event should be deprecated in favor of methods
that take the Event class as a parameter, but this will need an
audit to determine where it is necessary to use the event type name
(i.e. we do not have access to the Event object).

Closes #3319