Validate event disposition #3290
This commit validates that the event disposition value on the event
is a valid value within the chosen event disposition category.

This does not mean that there is any sort of semantic check of the
value if the disposition category changes or anything of the sort, this
is just a check that the value exists within the disposition category.

Closes #3290