Event is a required parameter #3319
This change removes the default parameter on these methods
to ensure that the activities returned are scoped to the given event.

Also, this commit removes an unused instruments method on
InstrumentPlan (which should really be renamed to
ActivityPlan or ParticipantActivityPlan or something like that),

All callers of these methods do send the event parameter so the
nil default could have been masking callers who had a nil event.