Revert nil check #3319
This commit is a partial revert of a036cfa6acb2be09d5c454699953eab4a84253f6
which changed the instrument_plan to expect an event
parameter without a nil default.  During testing the bug #3180 manifested
again when the instrument event association was nil.

The previous commit 085388cbf52ab35e3423f9efa57f0cee6f2926d4
addressed one instance where this was a problem, but until we
fix this association (as documented in the method comments)
in task #3341 I've re-inserted the nil check in the instrument plan.