Thread instrument mode through Instrument.start.
This ensures that an instrument's mode of contact is immediately present
after it's been saved.  Previously, an instrument's mode of contact
would only be recorded if

1. the instrument had a multi-part survey
2. a part following the first part was started

(The above behavior looks accidental.)

This commit doesn't try to straighten out the confusing nature of
Instrument.start and Person#start_instrument.  That has to be fixed
(there really needs to be separate methods for starting vs. continuing
instruments, since they're different tasks) but that needs to be done in
a separate commit.