Update to RSpec 2.12. #2744.
I want to be able to use parameterized shared contexts in my response
set populator tests.  (In my case, the parameter is the name of a test
survey that will be loaded in a before-all.)  This capability doesn't
exist in RSpec 2.6.

Said capability was probably introduced sometime before 2.12, but I've
been using 2.12 for a while and it seems stable enough, so might as well
just use it.