Filter consents by event date range. #3263.
N.b.: I'm not sure if the use of Date::Infinity here is kosher. It
works and all, but I can't tell if it is intended to be part of the
public API. It is listed in rubyspec, but only with stubs. And, as
noted in the code, it works as the end of a range but not the start
(since it doesn't implement #succ).

Anyway, future developers feel free to curse my memory when it gets
removed it in Ruby 3.14.