Shift lo-hi determination to informed consent events. Closes #3263.
This commit changes the import-time lo-hi determination decision in
two ways. First, It moves it out of OperationalImporter and into
`Participant#set_state_for_event_type` along with all the other state
detail decisions. Second, it associates the decision with each
Informed Consent event, instead of the Low-High Conversion event. (It
turns out that the Low-High conversion event is not consistently
recorded across SCs.)

This required a fair number of changes to the importer specs. Some are
expected and obvious reflections of the new behavior. Others are
because the new model resulted in the test participant's odd sequence
of events triggering a (pre-existing) implicit conversion to hi on
PV1.