Use SurveyTaker in ODE and RSP specs.  #2744.
The shim around SurveyTaker#respond in SurveyCompletion is quite a mess,
but it minimizes changes.  Further fixes welcome.

This commit also fixes a few errors in test surveys:

* RACE_1_OTH questions with answers whose response classes that didn't
  match the actual instruments
* incorrect reference identifiers in the birth visit surveys
* incorrect reference identifiers in the postnatal survey

Finally, this commit deletes four examples in the PregnancyScreener ODE
that deal with different date formats.  The examples were deleted
because date parsing is not the ODE's job; that's the job of whatever is
sending values into Surveyor's datastore, i.e. Response#value=.