Re-add call to mark_activity_occurred
Commit 2a537ce8a1ad4d0bf0b8e841478dfc0980927aae
added an mark_activity_occurred call from update.

The very next commit for this file
0fecc6c310b06ba188f4055a1a4458baf2887f67
removed this call.

This commit re-adds it. We need this call after update
because users can get stuck when performing activities
for an event at the informed consent activity for a participant
with an existing consent.

Changes in #3170 and specifically the task #3194 should
address the issue.