extracted eligibility prefix into method
The commit for birth cohort #3324 introduced the difference
between a PBS screening instrument and a Birth cohort
screening instrument.

Unfortunately other places in the code (SampledPersonsIneligibility)
called some methods whose signatures had been updated.

This commit removes the additional parameters to a method which
determines the instrument used to screen the participant.