Added Event.psc_ideal_date #3411
This commit adds the attribute psc_ideal_date to Event.
The psc_ideal_date replaces the event_start_date as one of
the coupling attributes to PSC (this along with event_type).

This change also updates the PSC matching criteria to use the
psc_ideal_date attribute instead. The migrations add the column to
the database as well as update all existing records to set the psc_ideal_date
to the current event_start_date.

Additionally there is a before_save callback to set the psc_ideal_date to that
of the event_start_date if that attribute is not set.