Prevent editing event_type #3412
This commit updates the event edit form to prevent the editing
of the event_type attribute. Also this change includes the ability
to update the event_start_date without any additional steps. #3411

I have kept/added the ability to update the psc_ideal_date but this
is only for system administrators. The reason behind this is to allow
sys_admins to easily update this attribute in order to correct any
broken PSC associations.

Closes #3412