common redirect logic refactored #3293
This ads top level routes for email and telephone to be consistent with
address and to ease the refactor. Makes redirect logic independent of
the URL which eases adding new redirects in the future eg. to provider page.
Tests added for the three redirect cases (no person, non-participant,
participant) for create and update.