Do not show previous reference activities
Reference activities are the non-first part of a multi-part
survey. When an activity of this sort has already been completed
it is not necessary to show this activity in the "previous_activity"
partial - since the response set associated with the activity will be
shown in the activity that this activity references.