consent_form_type should be CONSENT_TYPE_CL3
The MDES for 2.0 got the code list wrong for the consent_form_type
attribute on the ParticipantConsent. The MDES said CONSENT_TYPE_CL1
but should have used CONSENT_TYPE_CL3 (which unfortunately is not a
code list in the 2.0 MDES).

This commit adds the CONSENT_TYPE_CL3 code list to the MDES 2.0
ncs_codes and corrects the consent_form_type code list association.