Update to Sidekiq 2.8.0.
We were previously on Sidekiq 2.5.2.  However, the Sidekiq::Workers,
Sidekiq::Stats, and other critical APIs[1] were added in later versions.

I would like to be able to manage Sidekiq workers from the Rails
console, and I think others would like that same ability.  The 2.5.2 ->
2.8.0 change seems to have not broken Sidekiq jobs or related code (i.e.
code that uses Redis as a datastore), so this change seems safe to me.