Interpret pbs_birth_cohort flag during sync. Closes #3510.
This flag is not currently ever set by `Psc::SyncLoader`. This means
that eligibility events will always be sync'd using the PBS provider-
recruited PSC segment. In the future (#3512) we will need to determine
how to set this flag appropriately in `Psc::SyncLoader`. (For now we
are not supporting birth cohort syncing from Field.)