Make IC spec demonstrate response set requirement.  #2744.
Specifically, this change reduces the number of sites that instantiate
an InstrumentContext to one.  This single site instantiates an
InstrumentContext with a ResponseSet.

This change also makes a great number of changes to examples in the spec
to straighten out example flows.  It also corrects some obviously
incorrect example descriptions, i.e. those for age_of_child_in_months.

BIG IMPORTANT NOTE: There are likely incorrect behaviors in
InstrumentContext that are incorrectly supported by the examples.  One
major auditing task is that helpers in InstrumentContext should _not_
read responses associated with the ResponseSet's _person_ but rather the
_participant_.

I haven't tried to fix any of those issues.  Those will have to be fixed
in other commits.