Do not override method #3526
This override was necessary at one point, iirc, to
handle how surveyor was displaying datetime fields
in the UI. Surveyor 1.2 updated this method and there
is no longer a need to override. In fact, date and time
fields in Surveyor are broken in Cases due to this override
so we should not have this at all.

Overridden methods should be rare.