Correct schema version.
20130306200153 is the latest migration. It appears that the commits
that introduced 20130306200153 and 20130301164426 were merged in the
reverse order that they were created. The schema changes for both are
present, but the declared version is the earlier one.

The result is that `db:reset` puts the database in a state where the
changes from 20130306200153 have been applied but the migrations table
doesn't reflect it. This means that the tasks that won't run when
there are unapplied migrations (e.g., `db:test:prepare`) tell you to
run the migrations. But when you run the migrations 20130306200153
fails because the column it tries to add already exists.