Child Consent #3445
In allowing for multiple consents for a child participant
the way we create child participants needed to change.
Previously the child consent itself created the child participant record,
this commit de-couples the child person/participant record creation
from the child participant consent record adding a way to create/update
that child record on it's own.