Correct spec for real this time.
Participant#next_scheduled_event requires there to be at least one
contact. This commit modifies the spec so that that contact's event is
already completed. This means that the only pending events will be the
the results from #schedule_and_create_placeholder which the spec is
about, so the spec's expectations can be written relative to the
pending events.