Re-work logic for consented? #3560
For determining if the participant is consented?
reconsented? or withdrawn?

or consented to sample collection

we use the most recent ParticipantConsent record to
determine consent.

Missing from this commit is a test if the most recent
consent has sample consents and if this is not
true (and consent is given) to check the previous
consent's sample consent records.