Make code list cache work in development. #3633.
When config.cache_classes is false, the NcsCode Class instance that
was current when a code list was cached may not be the version in use
when the code list value is set to a model attribute.

When this happens, the `case ...; when NcsCode; ...` condition in
`NcsCodedAttributeValueHelpers#set_ncs_code` fails to match, resulting
in an exception with a very confusing message.

This commit rectifies this situation by clearing cached NcsCode
instances from an out-of-date NcsCode Class on first access, allowing
them to be reloaded from the database.