Check for participant.person
The recent commit e61f0bc1 changed the method used
to find participants to update. This change returns participants
without associated person records (how this is possible I do not
know and requires investigation).

This commit simply checks if this association exists and if not
then logs that record and continues (as the person association
is required to cancel activities in PSC as the person public_id
is the unique id in PSC.)