Continue to next event #3861
When continuing on to the next event in the same contact we were being
redirected improperly after the completion of the consent record as the event
was thought to be closed (the event in this case being the first event in the
contact).

This commit adds a method to determine the current event in case there is
more than one event for the contact.

This is a hot fix for 1.5.0.