Limit continuable action #3880
Do not show the continue action if the event is
an informed consent event that is associated with
other events.

Also do not show the continue action if the event
is closed (non consent events)

This is a temporary fix for #3880 - the permanent fix
should be a different solution but done at a later date.