Update to ncs_mdes_warehouse 0.13.1. Closes #4003.
Some locations which will be on 1.5.x need features that are new in
ncs_mdes_warehouse 0.13.x in their warehouse configurations. They need
to be on warehouse 0.13.x for the import scripts to be able to read
their import source warehouse configurations.

This change has already been made in `develop` and so will not be
ported forward.