Determine hi entirely based on consents given. #3997.
This change eliminates the requirement that a participant consent be
associated with consent event in order to be able to trigger a lo-hi
conversion. The import process will now check if the participant
should be in hi at each event based on her consent history. She will
be transitioned to hi for the first event which ends on or after her
first hi consent.