Eliminate continuable? confusion.
There were two helpers named `continuable?` with different signatures.
The one in ContactLinksHelper "won", which caused the spec in
`application_helper_spec` to fail.

It turns out that ApplicationHelper#continuable? was not used outside
of its own tests, so this commit removes it.

ContactLinksHelper#continuable? was not tested and not used outside
of the helper, so this commit makes it private.