Update based on feedback #3802
The Participant#date_available_for_informed_consent_event?
method should only take a Date (not either a Date or String) so
update this method and callers accordingly.

This date can be given by the user calling the rake task to
schedule reconsent events via highline.

In all places attempting to schedule a participant notify the
user of the progress or of any issues.