Ensure Answer associated with Question #4084
When creating or updating a Response validate that the
Answer being set is one of the Answers associated with the Question.

This probably should be added to Surveyor as a validation but is included
here to address the #4084 issue going forward with Cases. If Surveyor is
ever changed to add this validation we should remove it here
(see TODO note above validate callback)