Fix MDES loading from PG dump and yaml #4143
Non-null constraints were introduced at some point to the updated_at
column for the ncs_codes table. When this happened the associated
pg_dump file for MDES 3.2 no longer worked as it included NULL values
in the updated_at fields of the ncs_codes table.

This commit includes an updated pg_dump file that specifies non-null
values where necessary, and adds the CURRENT_TIMESTAMP values to the
updated_at column when the codes are loaded from a YAML file.