keep DOB consistent #3246
This will clear person_dob_date if perosn_dob is set to nil, unknown,
refused or a MDES partial date. The consequence is that the
operational enumerator will use the age and age_range instead of
computing it from the out-of-sync person_dob_date.

Displaying person_dob on the UI allows the user to see
e.g. '9666-96-96' when they set the DOB to unknown. It may
be better if we select the radio button and keep the DOB
blank. Although the current approach shows the actual data.