Use `before(:all)` init for subject in bulk of specs. #3996.
The slowest part of this suite before this commit was iterating over
all the various associations and verifying if they were cloned (or
not). This was slow because the clone itself is slow. But since the
test data doesn't change for this subset of the specs, the clone can
safely be computed once (in `before(:all)`) and then each association
examined in a separate spec.

On my laptop, this commit provides a 33x speedup for the full
case_cloner_spec (10 minutes down to to 18 seconds).