Remove required and add -4 option #4553
As part of hot fix 1.9.4 we were asked to remove the mandatory flag
on four fields on the Event form. In addition, to make it easier for data
collectors to understand the value for these not required fields we
re-added 'Missing in Error' as an option to the select list for code list
attributes.