fix check for invalid codes during ode #4535
This should solve a category of erros where the code list being
extracted to in operational data is slightly different than the code
list in instrument data.

There is an argument for not even extracting these code lists but this
fix simply prevents trying to extract invalid codes and is the minimal
change.

also see #4448