A terrible yet workable last-ditch solution for missing values.  #4571.
As we transition data out of NCS Navigator for one of our 1.6.7 sites,
we have found several situations in which instrument data was truly not
entered.  InstrumentToWarehouse was designed to fill in these missing
values with -4 (i.e. "Missing in Error"), but it does not accomplish
this in all cases.  Some examples:

* responses missing because one part of a multi-part instrument was
  never started
* responses missing due to some unknown cause (see e.g. [1] if you have
  access to NUBIC's internal issue tracker)

This horrifying, cheesy, yet effective solution makes a last pass over
the created MDES record, and just fills in -4s where it can.  It farts
in the general direction of data quality, but we have been told that
quality now takes a backseat to _just getting the data out_.  Also, it
works.

This commit has no tests because writing test cases for this sort of
thing takes longer than just demonstrating it works by running ETL jobs
on test instances.

[1]: https://code.bioinformatics.northwestern.edu/redmine/issues/show/9680