Disable response -> answer association check.
Once, we were trying to figure out why we were getting bad answer IDs
for questions.  We now know why (it's related to date entry in some old
surveys) and we have a way to fix corrupted data.

We do not, however, have a way to fix data that never made it into
Cases' database in the first place.  As such, it's better to take this
check out.